requirements.txt
django==4.0.1
django-extensions
envparse
psycopg2-binary
gunicorn
djangorestframework
social-auth-app-django
django-filter
pytest-django
pytest-factoryboy
pydantic